Set initial_npcs correctly in the showcase tool
initial_npcs is filled based on which npcs were selected to initially
spawn in various locations, and is used to decide whether to trigger
additional actions/dialogue in certain cases when reaching discordance.
particularly relevant here, is that Tomas will get spawned if you strike
with Isabella in her Community Rivals or The Exile events.